requirements.txt
antlr4-python3-runtime
antlr4-tools
black
cfpq-data
grammarinator @ git+https://github.com/renatahodovan/grammarinator.git@f3ffa71
networkx==3.2.1
pre-commit
pydot
pytest
pytest-xdist
scipy
